+++ src/main/kotlin/org/wfanet/measurement/duchy/deploy/common/server/BUILD.bazel
@@ -151,3 +151,30 @@ kt_jvm_library(
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/identity",
     ],
 )
+
+kt_jvm_library(
+    name = "forwarded_storage_postgres_duchy_data_server",
+    srcs = ["ForwardedStoragePostgresDuchyDataServer.kt"],
+    deps = [
+        "//src/main/kotlin/org/wfanet/measurement/duchy/deploy/common/server:duchy_data_server",
+        "//src/main/kotlin/org/wfanet/measurement/duchy/deploy/common/service:postgres_duchy_data_services",
+        "@wfa_common_jvm//imports/java/picocli",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/db/postgres:flags",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/db/r2dbc/postgres",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/storage/forwarded",
+    ],
+)
+
+java_binary(
+    name = "ForwardedStoragePostgresDuchyDataServer",
+    main_class = "org.wfanet.measurement.duchy.deploy.common.server.ForwardedStoragePostgresDuchyDataServerKt",
+    runtime_deps = [":forwarded_storage_postgres_duchy_data_server"],
+)
+
+java_image(
+    name = "forwarded_storage_postgres_duchy_data_server_image",
+    binary = ":ForwardedStoragePostgresDuchyDataServer",
+    main_class = "org.wfanet.measurement.duchy.deploy.common.server.ForwardedStoragePostgresDuchyDataServerKt",
+    visibility = ["//src:docker_image_deployment"],
+)
