+++ src/main/kotlin/org/wfanet/measurement/duchy/deploy/common/service/BUILD.bazel
@@ -19,8 +19,8 @@ kt_jvm_library(
     name = "postgres_duchy_data_services",
     srcs = ["PostgresDuchyDataServices.kt"],
     visibility = [
+        "//src/main/kotlin/org/wfanet/measurement/duchy/deploy/common/server:__pkg__",
         "//src/main/kotlin/org/wfanet/measurement/duchy/deploy/gcloud/server:__pkg__",
-        "//src/main/kotlin/org/wfanet/measurement/duchy/deploy/postgres/server:__pkg__",
         "//src/test/kotlin/org/wfanet/measurement/integration/common/duchy:__pkg__",
     ],
     deps = [
