+++ src/main/kotlin/org/wfanet/measurement/duchy/deploy/gcloud/server/BUILD.bazel
@@ -57,34 +57,6 @@ java_image(
 )
 
 kt_jvm_library(
-    name = "forwarded_storage_postgres_duchy_data_server",
-    srcs = ["ForwardedStoragePostgresDuchyDataServer.kt"],
-    deps = [
-        "//src/main/kotlin/org/wfanet/measurement/duchy/deploy/common/server:duchy_data_server",
-        "//src/main/kotlin/org/wfanet/measurement/duchy/deploy/common/service:postgres_duchy_data_services",
-        "//src/main/kotlin/org/wfanet/measurement/duchy/deploy/gcloud/spanner/computation",
-        "@wfa_common_jvm//imports/java/picocli",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/db/postgres:flags",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/db/r2dbc/postgres",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/storage/forwarded",
-    ],
-)
-
-java_binary(
-    name = "ForwardedStoragePostgresDuchyDataServer",
-    main_class = "org.wfanet.measurement.duchy.deploy.gcloud.server.ForwardedStoragePostgresDuchyDataServerKt",
-    runtime_deps = [":forwarded_storage_postgres_duchy_data_server"],
-)
-
-java_image(
-    name = "forwarded_storage_postgres_duchy_data_server_image",
-    binary = ":ForwardedStoragePostgresDuchyDataServer",
-    main_class = "org.wfanet.measurement.duchy.deploy.gcloud.server.ForwardedStoragePostgresDuchyDataServerKt",
-    visibility = ["//src:docker_image_deployment"],
-)
-
-kt_jvm_library(
     name = "gcs_spanner_computations_server",
     srcs = ["GcsSpannerComputationsServer.kt"],
     deps = [
