--- src/main/kotlin/org/wfanet/measurement/duchy/deploy/postgres/tools/BUILD.bazel
@@ -1,20 +1,0 @@
-load("@rules_java//java:defs.bzl", "java_binary")
-load("//src/main/docker:macros.bzl", "java_image")
-
-java_binary(
-    name = "UpdateSchema",
-    args = ["--changelog=duchy/postgres/changelog.yaml"],
-    main_class = "org.wfanet.measurement.common.db.postgres.tools.UpdateSchema",
-    resources = ["//src/main/resources/duchy/postgres"],
-    runtime_deps = [
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/db/postgres/tools:update_schema",
-    ],
-)
-
-java_image(
-    name = "update_schema_image",
-    args = ["--changelog=duchy/postgres/changelog.yaml"],
-    binary = ":UpdateSchema",
-    main_class = "org.wfanet.measurement.common.db.postgres.tools.UpdateSchema",
-    visibility = ["//src:docker_image_deployment"],
-)
